# internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go + PocketBase backend for QuantumAtlas: paper\ncollection + search + database. Write endpoints and\nthe search/paper read surface require a bearer token\n(PAT or PocketBase session) plus the matching scope.\nSee the auth model docs for the scope vocabulary.",
        "title": "QuantumAtlas API",
        "contact": {
            "name": "QuantumAtlas",
            "url": "https://quantum-atlas.ai"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/admin/assets/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only: find papers with at least one asset.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Search papers with assets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/admin/assets/{paper_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only asset listing: every PDF / markdown object the\npaper owns, with S3 object keys, sizes, SHA256s.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List paper assets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "paper id (qa_...)",
                        "name": "paper_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/assets/{paper_id}/{kind}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only proxy stream with Content-Disposition: attachment.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Download paper asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "paper id",
                        "name": "paper_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pdf | markdown",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/admin/assets/{paper_id}/{kind}/url": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only: time-limited presigned URL for direct browser-to-S3.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Presigned asset URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "paper id",
                        "name": "paper_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pdf | markdown",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "duration (default 1h, max 24h)",
                        "name": "ttl",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003ctoken\u003e\" — two credential shapes\naccepted: a Personal Access Token\n(`Authorization: Bearer qat_...`, minted at\n`/pat` after GitHub OAuth login), or the\nconfig-loaded system PAT (set\n`system_pat.token` in config.yaml on the\nserver, send the plaintext as\n`Authorization: Bearer \u003cvalue\u003e`).\nBrowser callers are authenticated through\npb.authStore (no copy step) — only non-browser\ncallers need an explicit bearer.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}